% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\name{create_option}
\alias{create_option}
\title{Create option}
\usage{
create_option(name, value)
}
\arguments{
\item{name}{(\link{character}) Option name}

\item{value}{(\link{character}) Option value}
}
\value{

}
\description{
Creates the desired option "directly" in the global options (as opposed to in
the environment created in the global options via
\link{init_package_options}).
}
\examples{
create_option("hello", "world")
getOption("hello")
}
