% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\name{get_option}
\alias{get_option}
\title{Get options (via names list)}
\usage{
get_option(name, env = getOption(pops:::option_env_name()))
}
\arguments{
\item{name}{(\href{1}{character}) Name to get}

\item{env}{(\link{environment}) Custom environment in options. See
\link{init_package_options}}
}
\value{

}
\description{
Get atomic option value from option environment created via
\link{init_package_options}.
}
\examples{
init_package_options(values = list(hello = "world", test = TRUE))
get_option("hello")
try(get_option())
# Same as
getOption(pops:::option_env_name())$hello
}
