% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\name{get_options}
\alias{get_options}
\title{Get options (via names list)}
\usage{
get_options(names = character(), env = getOption(pops:::option_env_name()))
}
\arguments{
\item{names}{(\href{1}{character}) Vector of names to get.}

\item{env}{(\link{environment}) Custom environment in options. See
\link{init_package_options}}
}
\value{

}
\description{
Get option values from option environment created via
\link{init_package_options}.
}
\examples{
init_package_options(values = list(hello = "world", test = TRUE))
get_options()
get_options(c("hello", "test"))
get_options(c("hello"))
}
