% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops-package.R
\docType{package}
\name{pops-package}
\alias{pops}
\alias{pops-package}
\title{pops: Options management for R packages}
\description{
Initializing, getting and setting of package options.
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
