% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\name{set_options}
\alias{set_options}
\title{Set options (via names list)}
\usage{
set_options(values = list(), env = getOption(pops:::option_env_name()))
}
\arguments{
\item{values}{(\link{list}) Named list with name-value pairs}

\item{env}{(\link{environment}) Custom environment in options. See
\link{init_package_options}}
}
\value{

}
\description{
By default, the option is created \strong{inside} the environment that is
created/set as an option via \link{init_package_options} (see default
value of the \code{env} argument).
}
\examples{
init_package_options()
set_options(list(hello = "world", test = TRUE))
get_options()
get_option("hello")
# Same as
getOption(pops:::option_env_name())$hello
}
